Basic Functionality Implementation
Added the basic "Controller" buttons and a default "grid-style" drawing
area. Will be converted into a "canvas-style" container in a future
commit. Also implemented "mousedown" check to make it more accessible.

In future commits:

- Canvas (instead of grid),
- Rainbow mode,
- The grid slider implementation (instead of prompting the user),
- A working "clear" button that actually _does clear_ the *background*
of the grid divs instead of removing them altogether,
- Further styling for an elegan(ish) look,
- Possibly syntactic sugaring here and there,

are to be implemented, which will be Soon™.