Finalized UI and Functionality
Both the UI and the functionality has been finalized as of this first
fihnished version. There will be updates in the future to polish it even
more; but for now I am shelving this little project to finally continue
where I had left off a long time ago.

Changes made in this commit:

- Fixed the Rainbow mode,
- Implemented minor logical changes to the way the code works,
- Added some UI improvements to make it look a little bit nice,
- Added the credit footer,
- Fixed the non-working features from previous commits

Things to improve on (as they pop on my mind as of right now):

- A theme selection feature,
- More responsible CSS work,
- Some extra features like fill and shade etc. that I've seen from other
projects under this lesson.
- Further miscellaneous code changes and fixes if needed.